+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Compress Screenshots on a background thread ([#4295](https://github.com/getsentry/sentry-java/pull/4295))
 - Improve low memory breadcrumb capturing ([#4325](https://github.com/getsentry/sentry-java/pull/4325))
+- Make `SystemEventsBreadcrumbsIntegration` faster ([#4330](https://github.com/getsentry/sentry-java/pull/4330))
 
 ## 7.22.5